# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Gearbox API",
        "description": "REST API untuk Sistem Booking & Manajemen Layanan Bengkel Mobil",
        "contact": {
            "email": "admin@gearbox.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "Local Development Server"
        }
    ],
    "paths": {
...
        "/api/schedules": {
            "get": {
                "tags": [
                    "Schedules"
                ],
                "summary": "Get semua jadwal layanan",
                "operationId": "f21805e74c0fcc7937d0464ff67212bd",
                "parameters": [
                    {
                        "name": "service_id",
                        "in": "query",
                        "description": "Filter berdasarkan ID layanan",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    },
                    {
                        "name": "date",
                        "in": "query",
                        "description": "Filter berdasarkan tanggal (format: Y-m-d)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "2026-05-25"
                        }
                    },
                    {
                        "name": "is_available",
                        "in": "query",
                        "description": "Filter berdasarkan ketersediaan slot",
                        "required": false,
                        "schema": {
                            "type": "boolean",
                            "example": true
                        }
                    },
                    {
                        "name": "limit",
                        "in": "query",
                        "description": "Jumlah data per halaman",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 10
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List jadwal dengan pagination",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "current_page": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "service_id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "date": {
                                                        "type": "string",
                                                        "example": "2026-05-25"
                                                    },
                                                    "start_time": {
                                                        "type": "string",
                                                        "example": "08:00:00"
                                                    },
                                                    "end_time": {
                                                        "type": "string",
                                                        "example": "10:00:00"
                                                    },
                                                    "capacity": {
                                                        "type": "integer",
                                                        "example": 3
                                                    },
                                                    "booked_count": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "is_available": {
                                                        "type": "boolean",
                                                        "example": true
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        },
                                        "total": {
                                            "type": "integer",
                                            "example": 20
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/schedules/{id}": {
            "get": {
                "tags": [
                    "Schedules"
                ],
                "summary": "Get detail jadwal layanan",
                "operationId": "0d2db25e1750eaff76142e434442ee13",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID jadwal",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detail jadwal",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "service_id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "date": {
                                            "type": "string",
                                            "example": "2026-05-25"
                                        },
                                        "start_time": {
                                            "type": "string",
                                            "example": "08:00:00"
                                        },
                                        "end_time": {
                                            "type": "string",
                                            "example": "10:00:00"
                                        },
                                        "capacity": {
                                            "type": "integer",
                                            "example": 3
                                        },
                                        "booked_count": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "is_available": {
                                            "type": "boolean",
                                            "example": true
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Jadwal tidak ditemukan"
                    }
                }
            }
        },
        "/api/admin/schedules": {
            "post": {
                "tags": [
                    "Admin"
                ],
                "summary": "Tambah jadwal baru (Admin)",
                "operationId": "37ab2fe4be094791d04c5a0df9cf7f18",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "service_id",
                                    "date",
                                    "start_time",
                                    "end_time",
                                    "capacity"
                                ],
                                "properties": {
                                    "service_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "date": {
                                        "description": "Format: Y-m-d, tidak boleh lebih awal dari hari ini",
                                        "type": "string",
                                        "example": "2026-06-01"
                                    },
                                    "start_time": {
                                        "description": "Format: H:i",
                                        "type": "string",
                                        "example": "08:00"
                                    },
                                    "end_time": {
                                        "description": "Format: H:i, harus setelah start_time",
                                        "type": "string",
                                        "example": "10:00"
                                    },
                                    "capacity": {
                                        "description": "Maksimal booking per slot",
                                        "type": "integer",
                                        "example": 3
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Jadwal berhasil ditambahkan",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Jadwal berhasil ditambahkan"
                                        },
                                        "schedule": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Unauthorized — bukan admin"
                    },
                    "422": {
                        "description": "Jadwal duplikat atau validasi gagal"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/admin/schedules/{id}": {
            "put": {
                "tags": [
                    "Admin"
                ],
                "summary": "Update jadwal (Admin)",
                "operationId": "e4c841501d2501cb6315fdb08933759d",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID jadwal",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "date": {
                                        "type": "string",
                                        "example": "2026-06-02"
                                    },
                                    "start_time": {
                                        "type": "string",
                                        "example": "09:00"
                                    },
                                    "end_time": {
                                        "type": "string",
                                        "example": "11:00"
                                    },
                                    "capacity": {
                                        "type": "integer",
                                        "example": 5
                                    },
                                    "is_available": {
                                        "type": "boolean",
                                        "example": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Jadwal berhasil diupdate"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Jadwal tidak ditemukan"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Admin"
                ],
                "summary": "Hapus jadwal (Admin)",
                "operationId": "83b3381746d7de0804af2255e9f5432c",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID jadwal",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Jadwal berhasil dihapus"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Unauthorized"
                    },
                    "422": {
                        "description": "Tidak bisa hapus — sudah ada booking aktif"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Masukkan token dari response login",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}